+++ backend/gumison/src/main/java/com/ssafy/gumison/api/service/HistoryService.java
@@ -19,6 +19,6 @@ public interface HistoryService {
   public List<Solution> createSolution(SolutionRequest solutionRequest);
 
   public List<Solution> updateSolution(SolutionRequest solutionRequest);
-  
-  public Long deleteSolution(String uploadId);
+
+  public Integer deleteSolution(String uploadId);
 }
